Reorg file structure so <App /> is testable
